NCFE-67: Separate questions from functions
